// src/models/IoK8sApiAdmissionregistrationV1ValidatingWebhookConfigurationList/types.ts
import { IoK8sApiAdmissionregistrationV1ValidatingWebhookConfigurationModelOptions as IoK8sApiAdmissionregistrationV1ValidatingWebhookConfiguration } from '../IoK8sApiAdmissionregistrationV1ValidatingWebhookConfiguration';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiAdmissionregistrationV1ValidatingWebhookConfigurationListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiAdmissionregistrationV1ValidatingWebhookConfiguration>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}